Add polybar toggler key.

+++ polybar/dot-config/polybar/launch.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
 
-pkill polybar
-
-if type "xrandr"; then
-    for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
-        MONITOR=$m polybar --reload main &
-    done
-else
-    polybar --reload main &
+if ! pkill polybar; then
+    pkill polybar
+    if type "xrandr"; then
+        for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
+            MONITOR=$m polybar --reload main &
+        done
+    else
+        polybar --reload main &
+    fi
 fi
-
